* '4.7' of scm.dev.nokia.troll.no:qt/oslo-staging-2:
  Fixed potential crash when loading corrupt GIFs.
  Work around an ATI driver problem with mutli-sampled pbuffers.
  tst_qstatemachine.cpp: fix compilation with Sun Studio
  Fixed regression in clipping.qps autotest on 64-bit.
  Fixed crash when using Qt::WA_DeleteOnClose on a QPrintDialog on Mac.
  Fixed performance regression in curve stroking.
  Don't disable texture_from_pixmap on GLX/X11 by default.
  Avoid creating copy of an image in memory when storing as png
  Doc update for the support of MSVC 2010 64-bit
  fix documentation of drawText(int, int, int, int, ...
  Optimization of pixel conversion when storing jpeg
  Don't pretend to support single buffered EGL surfaces.
  Named anonymous struct in the OpenGL paint engine.
  Fixes gray_raster incorrectly reporting out of memory error.

+++ src/gui/painting/qrasterdefs_p.h
@@ -100,7 +100,7 @@ QT_FT_BEGIN_HEADER
   /*    distances in integer font units, or 16,16, or 26.6 fixed float     */
   /*    pixel coordinates.                                                 */
   /*                                                                       */
-  typedef signed long  QT_FT_Pos;
+  typedef signed int  QT_FT_Pos;
 
 
   /*************************************************************************/
